Update scraper.py (Fix + feature)
Fix scraping where it couldn't find div[role="article"].
This div no longer exists now, so I fixed it.

Feature:
Add NUM_OF_PLACES in constants.json.
If it's not found it defaults to 10.

Update result name to include a timestamp at the end.
For more exclusive naming. It used to only save to the minutes, which
could lead to problems if 2 results came out in the same minute.